--- SiVL19/coco_caption/pycocoevalcap/meteor/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'tylin'
