--- SiVL19/coco_caption/pycocoevalcap/tokenizer/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'hfang'
